+++ 2008/2008_0205__ga__primary__state__president.csv
@@ -1,0 +1,20 @@
+office,candidate,party,county,votes
+President of the United States,Rudy Giuliani,R,N/A,7162
+President of the United States,Mike Huckabee,R,N/A,326874
+President of the United States,Duncan Hunter,R,N/A,755
+President of the United States,Alan Keyes,R,N/A,1458
+President of the United States,John McCain,R,N/A,304751
+President of the United States,Ron Paul,R,N/A,28096
+President of the United States,Mitt Romney,R,N/A,290707
+President of the United States,Tom Tancredo,R,N/A,324
+President of the United States,Fred Thompson,R,N/A,3414
+President of the United States,Total Republican,R,N/A,963541
+President of the United States,Joe Biden,D,N/A,2538
+President of the United States,Hillary Clinton,D,N/A,330026
+President of the United States,Chris Dodd,D,N/A,904
+President of the United States,John Edwards,D,N/A,18209
+President of the United States,Mike Gravel,D,N/A,952
+President of the United States,Dennis J. Kucinich,D,N/A,2096
+President of the United States,Barack Obama,D,N/A,704247
+President of the United States,Bill Richardson,D,N/A,1879
+President of the United States,Total Democrat,D,N/A,1060851
